refactor: Move logic nodes to individual files
Implements ML-597
