+++ nodes/logic.py
@@ -1,173 +1,9 @@
-import torch
-from comfy_execution.graph import ExecutionBlocker  # type: ignore
 from comfy_execution.graph_utils import GraphBuilder, is_link  # type: ignore
 
-from .categories import LABS_CAT, LOGIC_CAT
+from .categories import LABS_CAT
 from .shared import ByPassTypeTuple, any_type
 
 MAX_FLOW_NUM = 10
-
-
-class Switch:
-    """Switches between two input values based on a boolean condition.
-
-    A logic gate that selects between two inputs of any type based on a boolean condition. When the
-    condition is True, it returns the 'true' value; otherwise, it returns the 'false' value. This node
-    is useful for creating conditional workflows and dynamic value selection.
-
-    Args:
-        condition (bool): The boolean condition that determines which value to return.
-            Defaults to False if not provided.
-        on_true (Any): The value to return when the condition is True. Can be of any type.
-        on_false (Any): The value to return when the condition is False. Can be of any type.
-
-    Returns:
-        tuple[Any]: A single-element tuple containing either the 'true' or 'false' value based on
-            the condition.
-
-    Notes:
-        - The node accepts inputs of any type, making it versatile for different data types
-        - Both 'on_true' and 'on_false' values must be provided
-        - The condition is automatically cast to boolean, with None being treated as False
-    """
-
-    @classmethod
-    def INPUT_TYPES(cls):  # type: ignore
-        return {
-            "required": {
-                "on_true": (any_type, {"lazy": True}),
-                "on_false": (any_type, {"lazy": True}),
-                "condition": ("BOOLEAN", {"default": True}),
-            }
-        }
-
-    RETURN_TYPES = (any_type,)
-    RETURN_NAMES = ("output",)
-    FUNCTION = "execute"
-    CATEGORY = LOGIC_CAT
-
-    def check_lazy_status(self, condition: bool, on_true: any = None, on_false: any = None) -> any:
-        if condition and on_true is None:
-            return ["on_true"]
-        if not condition and on_false is None:
-            return ["on_false"]
-        return None
-
-    def execute(self, on_true: any, on_false: any, condition: bool = True) -> tuple[any]:
-        return (on_true if condition else on_false,)
-
-
-class Blocker:
-    """Controls flow execution based on a boolean condition.
-
-    A utility node that blocks or allows execution flow based on a boolean flag. When the continue
-    flag is False, it blocks execution by returning an ExecutionBlocker. When True, it passes through
-    the input value unchanged.
-
-    Args:
-        continue (bool): Flag to control execution flow. When False, blocks execution.
-        in (Any): The input value to pass through when execution is allowed.
-
-    Returns:
-        tuple[Any]: A single-element tuple containing either:
-            - The input value if continue is True
-            - An ExecutionBlocker if continue is False
-
-    Notes:
-        - Useful for conditional workflow execution
-        - Can be used to create branches in execution flow
-        - The ExecutionBlocker prevents downstream nodes from executing
-    """
-
-    @classmethod
-    def INPUT_TYPES(cls):
-        return {
-            "required": {
-                "should_continue": ("BOOLEAN", {"default": False}),
-                "input": (any_type, {"default": None}),
-            },
-        }
-
-    RETURN_TYPES = (any_type,)
-    RETURN_NAMES = ("out",)
-    CATEGORY = LABS_CAT
-    FUNCTION = "execute"
-
-    def execute(self, should_continue: bool = False, input: any = None) -> tuple[any]:
-        return (input if should_continue else ExecutionBlocker(None),)
-
-
-class Compare:
-    """Compares two input values based on a specified comparison operation.
-
-    A logic gate that evaluates a comparison between two inputs of any type. The comparison is determined
-    by the specified operation, which can include equality, inequality, and relational comparisons. This
-    node is useful for implementing conditional logic based on the relationship between two values.
-
-    Args:
-        a (Any): The first value to compare. Can be of any type.
-        b (Any): The second value to compare. Can be of any type.
-        comparison (str): The comparison operation to perform. Defaults to "a == b".
-            Available options include:
-            - "a == b": Checks if a is equal to b.
-            - "a != b": Checks if a is not equal to b.
-            - "a < b": Checks if a is less than b.
-            - "a > b": Checks if a is greater than b.
-            - "a <= b": Checks if a is less than or equal to b.
-            - "a >= b": Checks if a is greater than or equal to b.
-
-    Returns:
-        tuple[bool]: A single-element tuple containing the result of the comparison as a boolean value.
-
-    Notes:
-        - The node accepts inputs of any type, making it versatile for different data types.
-        - If the inputs are tensors, lists, or tuples,
-          the comparison will be evaluated based on their shapes or lengths.
-        - The output will be cast to a boolean value.
-    """
-
-    COMPARE_FUNCTIONS = {
-        "a == b": lambda a, b: a == b,
-        "a != b": lambda a, b: a != b,
-        "a < b": lambda a, b: a < b,
-        "a > b": lambda a, b: a > b,
-        "a <= b": lambda a, b: a <= b,
-        "a >= b": lambda a, b: a >= b,
-    }
-
-    @classmethod
-    def INPUT_TYPES(cls):  # type: ignore
-        compare_functions = list(cls.COMPARE_FUNCTIONS.keys())
-        return {
-            "required": {
-                "a": (any_type,),
-                "b": (any_type,),
-                "comparison": (compare_functions, {"default": "a == b"}),
-            }
-        }
-
-    RETURN_TYPES = ("BOOLEAN",)
-    RETURN_NAMES = ("result",)
-    FUNCTION = "execute"
-    CATEGORY = LOGIC_CAT
-
-    def execute(self, a: any, b: any, comparison: str = "a == b") -> tuple[bool]:
-        try:
-            output = self.COMPARE_FUNCTIONS[comparison](a, b)
-        except Exception as e:
-            if isinstance(a, torch.Tensor) and isinstance(b, torch.Tensor):
-                output = self.COMPARE_FUNCTIONS[comparison](a.shape, b.shape)
-            elif isinstance(a, (list, tuple)) and isinstance(b, (list, tuple)):
-                output = self.COMPARE_FUNCTIONS[comparison](len(a), len(b))
-            else:
-                raise e
-
-        if isinstance(output, torch.Tensor):
-            output = output.all().item()
-        elif isinstance(output, (list, tuple)):
-            output = all(output)
-
-        return (bool(output),)
 
 
 class LoopStart:
